requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-google-genai
sentence-transformers
faiss-cpu
python-dotenv
pypdf
pinecone
